todo-list-django/src/requirements.txt
certifi==2016.2.28
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.3
psycopg2==2.7.3.2
pytz==2017.3
wincertstore==0.2
gunicorn
